requirements.txt
spotipy==2.23.0
requests==2.31.0
python-dotenv==1.0.0
PyQt6==6.6.1
cryptography==42.0.2 